% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gb_body}
\alias{gb_body}
\title{The ggplot annotate polygon layers that draw the GB body}
\format{
A list (elements are scaled so the screen bottom left is close to x=0, y=0)
}
\usage{
gb_body
}
\description{
The ggplot annotate polygon layers that draw the GB body
}
\keyword{datasets}
