% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggboy-demo.R
\name{ggboy_demo}
\alias{ggboy_demo}
\title{Game Boy image simulator}
\usage{
ggboy_demo(
  file,
  res = 160,
  crop = TRUE,
  ncols = 4,
  cols = c("#0f380f", "#306230", "#8bac0f", "#9bbc0f"),
  aggregator = mean
)
}
\arguments{
\item{file}{Path to image file or array returned by \code{png::readPNG()} or \code{read::JPEG{}}}

\item{res}{Horizontal resolution of output (default = 160). Vertical resolution will be computed to match the output aspect ratio.
Set to NULL for original image horizontal resolution.}

\item{crop}{Should the image be cropped to match the original Game Boy screen aspect ratio (default = TRUE)}

\item{ncols}{The number of 'greenscale' colours to use (default = 4, the original Game Boy specification)}

\item{cols}{The colours used in the fill scale (passed to \code{ggplot2::scale_fill_gradientn()}) defaults to an approximation of the original Game Boy shades of greem}

\item{aggregator}{Function used to aggregate pixel intensities for final image.
This can be any function (can be anonymous) that takes a vector and returns a single value (default = mean).}
}
\value{
A ggplot2 plot
}
\description{
Experimental - not used, not exported
}
\details{
\code{ggboy_demo()} shows the original (in greyscale), the cropped (in greyscale) and the Game Boy version of the image
}
