% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggboy-gif.R
\name{ggboy_gif}
\alias{ggboy_gif}
\title{Create a ggboy animated gif}
\usage{
ggboy_gif(
  gif,
  width = 160,
  height = 144,
  scale = 1,
  fps = 5,
  delay = NULL,
  frame_data = NULL,
  first_n = NULL,
  ...
)
}
\arguments{
\item{gif}{File path or URL of GIF}

\item{width}{output width (default 160 * scale)}

\item{height}{output height (default 144 * scale)}

\item{scale}{multiplicative scaling factor for width and height (default = 1)}

\item{fps}{frames per second. Ignored if delay is not NULL}

\item{delay}{delay after each frame, in 1/100 seconds. Must be length 1, or number of frames. If specified, then fps is ignored.}

\item{frame_data}{a dataframe of x, y and z values as returned by \code{image_to_frame()} (default = NULL)}

\item{first_n}{only use frames \code{1:first_n} of the gif. Useful for testing settings on a smaller subset of the gif}

\item{...}{named arguments passed on to \code{ggboy()}}
}
\description{
Create a ggboy animated gif
}
