% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggboy_save.R
\name{ggboy_save}
\alias{ggboy_save}
\title{Save a ggboy image with the correct dimensions}
\usage{
ggboy_save(filename, scale = 1)
}
\arguments{
\item{filename}{file name to create on disk}

\item{scale}{multiplicative scaling factor}
}
\description{
ggboy_save() is a convenient function for saving a ggboy image at the correct aspect ratio.
By default it saves images at 160 px (wide) x 144 px (high)
}
