% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hk}
\alias{hk}
\title{Photograph of Hong Kong by night}
\format{
An array as returned from \code{jpeg::readJPEG()}. Dimensions (height x width x channels) = 401 604 3
}
\usage{
hk
}
\description{
A photograph I took whilst on holiday in Hong Kong (2008)
}
\keyword{datasets}
