% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_to_frame.R
\name{image_to_frame}
\alias{image_to_frame}
\title{Create a frame for ggboy images from another image}
\usage{
image_to_frame(img, resize_filter = "Mitchell", gravity = "center")
}
\arguments{
\item{img}{path to image to use for frame}

\item{resize_filter}{resize filter to be used \code{magick::filter_types} (default = 'Mitchell')}

\item{gravity}{position of crop window relative to image (default = "center"). A string value from \code{magick::gravity_types()}. Values such as \code{"south"}, \code{"west"}, \code{"northeast"}}
}
\description{
Images will be scaled and cropped to 160 px (wide) x 144 px (high) (if not already these dimensions)
Then converted to greyscale with the outer 16 pixels border selected
}
