% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rstats}
\alias{rstats}
\title{A version of the R logo}
\format{
An array as returned from \code{png::readPNG()}. Dimensions (height x width x channels) = 300 300 4
}
\usage{
rstats
}
\description{
A version of the R logo
}
\keyword{datasets}
