% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggboy.R
\name{ggboy}
\alias{ggboy}
\title{Game Boy image simulator}
\usage{
ggboy(
  file,
  res = 160,
  crop = TRUE,
  ncols = 4,
  cols = c("#0f380f", "#306230", "#8bac0f", "#9bbc0f"),
  aggregator = mean,
  graphic = FALSE
)
}
\arguments{
\item{file}{Path to image file or array returned by \code{png::readPNG()} or \code{read::JPEG{}}.}

\item{res}{Horizontal resolution of output (default = 160). Vertical resolution will be computed to match the output aspect ratio.
Set to NULL for original image horizontal resolution.}

\item{crop}{Should the image be cropped to match the original Game Boy screen aspect ratio (default = TRUE).}

\item{ncols}{The number of 'greenscale' colours to use (default = 4, the original Game Boy specification).}

\item{cols}{The colours used to create a fill gradient (passed to \code{ggplot2::scale_fill_gradientn()})
defaults to an approximation of the original Game Boy shades of green \code{c("#0f380f", "#306230", "#8bac0f", "#9bbc0f")}.}

\item{aggregator}{Function used to aggregate pixel intensities for final image.
This can be any function (can be anonymous) that takes a vector and returns a single value (default = mean).}

\item{graphic}{Should an approximation of the Game Boy body be included with the plot? (default = FALSE). The graphic can only be shown
if crop is also set to TRUE}
}
\value{
A ggplot2 plot
}
\description{
\code{ggboy()} attempts to simulate the way an image would look if displayed on the original (1989) Game Boy screen.
By default, images are cropped (centerally) to maintain the original Game Boy screen aspect ratio and pixel numbers (height/width = 144/160).
}
